Change to `instance_review_candidate` view
Now, candidates are selected based on the TOTAL number of reviews and
discards, rather than just reviews. This means that an instance that
has been reviewed by one person and discarded by another will no longer
appear to other users. This change was implemented to speed up the
review process for when there are fewer users. At the same time, this
change increases the probability of conflicts occuring.